requirements.txt
opencv-python
ultralytics
ncnn
psycopg2
python-dotenv